bidding-client/pyproject.toml
[project]
name = "aptos-a2a-bidding-system"
version = "0.1.0"
description = "A2A-Aptos MVP Bidding System with Personal Agent and Service Agent"
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "aptos-sdk>=0.8.0",
    "PyYAML>=6.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]